In-game menu, NPC sprites, Misc
Added a second, in-game menu and hooked the menus with the Time_Manager. Added Kaitlin's NPC sprites (the NPCs only have the south-facing sprites for now) and modified the sizes and collision accordingly. Made it so roofs and intrusive UI stay disabled until the game starts, when they are enabled by CameraSwitcher and PlayerController respectively.